Only list the user's updates that are still actionable
Today when listing the user's updates we also include updates that
are part of releases that have been archived or disabled.
With this commit only the updates for releases that the user can still
influence will be shown.

The commit also adds a link to ask that all releases be included in the
updates listed if people want to see the old information.

Fixes https://github.com/fedora-infra/bodhi/issues/752

Signed-off-by: Pierre-Yves Chibon <pingou@pingoured.fr>